requirements.txt
future==0.17.1
imageio==2.5.0
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
numpy==1.17.2
opencv-contrib-python==4.1.1.26
OpenEXR==1.3.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycparser==2.19
PyEXR==0.3.7
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.4
tensorboardX==1.8
terminado==0.8.2
testpath==0.4.2
torch
torchvision
tqdm==4.35.0
openEXR
pyexr
multiprocessing-logging
ffmpeg-python